backend/pyproject.toml
[project]
name = "sleep-recommendation-system"
version = "0.1.0"
description = "Symphony AI Sleep Music Recommendation System with LangGraph Multi-Agent Workflow"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-restful>=0.3.10",
    "flask-restx>=1.3.0",
    "flask-cors>=4.0.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "aiofiles>=24.0.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
cpu = [
  "torch>=2.7.0",
  "torchvision>=0.22.0",
  "torchaudio>=2.7.0",
]
cu118 = [
  "torch>=2.7.0",
  "torchvision>=0.22.0",
  "torchaudio>=2.7.0",
]
cu126 = [
  "torch>=2.7.0",
  "torchvision>=0.22.0",
  "torchaudio>=2.7.0",
]
cu128 = [
  "torch>=2.7.0",
  "torchvision>=0.22.0",
  "torchaudio>=2.7.0",
]

[tool.uv]
conflicts = [
  [ { extra = "cpu"   }, { extra = "cu118" } ],
  [ { extra = "cpu"   }, { extra = "cu126" } ],
  [ { extra = "cpu"   }, { extra = "cu128" } ],
  [ { extra = "cu118" }, { extra = "cu126" } ],
  [ { extra = "cu118" }, { extra = "cu128" } ],
  [ { extra = "cu126" }, { extra = "cu128" } ],
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[[tool.uv.index]]
name     = "pytorch-cpu"
url      = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name     = "pytorch-cu118"
url      = "https://download.pytorch.org/whl/cu118"
explicit = true

[[tool.uv.index]]
name     = "pytorch-cu126"
url      = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name     = "pytorch-cu128"
url      = "https://download.pytorch.org/whl/cu128"
explicit = true

# 2) Tell uv where to look for each PyTorch package
[tool.uv.sources]
torch = [
  { index = "pytorch-cpu",   extra = "cpu"   },
  { index = "pytorch-cu118", extra = "cu118" },
  { index = "pytorch-cu126", extra = "cu126" },
  { index = "pytorch-cu128", extra = "cu128" },
]
torchvision = [
  { index = "pytorch-cpu",   extra = "cpu"   },
  { index = "pytorch-cu118", extra = "cu118" },
  { index = "pytorch-cu126", extra = "cu126" },
  { index = "pytorch-cu128", extra = "cu128" },
]
torchaudio = [
  { index = "pytorch-cpu",   extra = "cpu"   },
  { index = "pytorch-cu118", extra = "cu118" },
  { index = "pytorch-cu126", extra = "cu126" },
  { index = "pytorch-cu128", extra = "cu128" },
]
